@@ -19,7 +19,7 @@ version: 1.1.0+1
 
 environment:
   # sdk: ">=2.17.0 <3.0.0"
-  sdk: ">=2.19.0 <4.0.0"
+  sdk: ">=3.2.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -33,6 +33,7 @@ dependencies:
   intl: ^0.19.0
   equatable: ^2.0.3
   flutter_bloc: ^8.1.3
+  bloc_concurrency: ^0.2.3
   firebase_core: ^2.5.0
   firebase_auth: ^4.2.6
   cloud_firestore: ^4.3.2
@@ -55,7 +56,7 @@ dependencies:
   data_tables: ^1.0.3
   grouped_list: ^5.1.2
   flutter_native_splash: ^2.3.9
-  fluttertoast: ^8.1.2
+  fluttertoast: ^8.1.4
   file_picker: ^6.1.1
   enum_to_string: ^2.0.1
 
