@@ -101,14 +101,13 @@ class _SquadMatchupsPage extends State<SquadMatchupsPage> {
       SizedBox(height: 10),
     ];
 
-    m.coachMatchups.forEach((m) => {
-          matchupWidgets.add(MatchupCoachWidget(
+    m.coachMatchups.forEach((m) => matchupWidgets.add(MatchupCoachWidget(
             tournament: _tournament,
             authUser: _authUser,
             matchup: m,
             refreshState: widget.refreshState,
           ))
-        });
+        );
 
     return Expanded(
         child: ListView(
