@@ -60,7 +60,7 @@ class CheckboxFormField extends FormField<bool> {
                     ? Builder(
                         builder: (BuildContext context) => Text(
                           state.errorText!,
-                          style: TextStyle(color: Theme.of(context).errorColor),
+                          style: TextStyle(color: Theme.of(context).colorScheme.error),
                         ),
                       )
                     : null,
